BACKEND/requirements.txt
fastapi 
uvicorn[standard] 
sqlalchemy 
psycopg2-binary 
pydantic[email] 
python-dotenv 
python-multipart 
requests 
fastapi-mail 
openai 
groq 
PyJWT 
bcrypt